+++ src/axom/sidre/core/MFEMSidreDataCollection.cpp
@@ -8,8 +8,9 @@
 #ifdef AXOM_USE_MFEM
 
   #include <string>
-  #include <iomanip>  // for setw, setfill
-  #include <cstdio>   // for snprintf()
+  #include <iomanip>      // for setw, setfill
+  #include <cstdio>       // for snprintf()
+  #include <type_traits>  // for checking layout
 
   #include "conduit_blueprint.hpp"
 
@@ -509,13 +510,38 @@ void MFEMSidreDataCollection::createMeshBlueprintAdjacencies(bool hasBP)
     }
   }
 
+  int dim = mesh->SpaceDimension();
+
+  mfem::Array<int> tmp_verts;
+  int tmp_edge;
+  int tmp_face;
+  int tmp_orientation;
+
   for(int gi = 1; gi < pmesh->GetNGroups(); ++gi)
   {
+    // MFEM will build groups with zero shared vertices but still some shared
+    // edges or faces, so we need to check all possible elements
+
+    // Shared element views are not created if they would be empty though,
+    // as this causes issues during a reload
     int num_gneighbors = pmesh->gtopo.GetGroupSize(gi);
     int num_gvertices = pmesh->GroupNVertices(gi);
+    int num_gedges = pmesh->GroupNEdges(gi);
+    int num_gtris = pmesh->GroupNTriangles(gi);
+    int num_gquads = pmesh->GroupNQuadrilaterals(gi);
 
-    // Skip creation of empty groups
-    if(num_gneighbors > 1 && num_gvertices > 0)
+    bool has_shared_elements = num_gvertices > 0;
+    if(dim >= 2)
+    {
+      has_shared_elements |= num_gedges > 0;
+      if(dim >= 3)
+      {
+        has_shared_elements |= num_gtris > 0;
+        has_shared_elements |= num_gquads > 0;
+      }
+    }
+
+    if(has_shared_elements && (num_gneighbors > 1))
     {
       std::snprintf(group_str,
                     GRP_SZ,
@@ -545,13 +571,75 @@ void MFEMSidreDataCollection::createMeshBlueprintAdjacencies(bool hasBP)
         }
       }
 
-      sidre::View* gvertices_view =
-        group_grp->createViewAndAllocate("values", sidre::INT_ID, num_gvertices);
-      int* gvertices_data = gvertices_view->getData<int*>();
-
-      for(int vi = 0; vi < num_gvertices; ++vi)
+      if(num_gvertices > 0)
       {
-        gvertices_data[vi] = pmesh->GroupVertex(gi, vi);
+        sidre::View* gvertices_view =
+          group_grp->createViewAndAllocate("values", sidre::INT_ID, num_gvertices);
+        int* gvertices_data = gvertices_view->getData<int*>();
+
+        for(int vi = 0; vi < num_gvertices; ++vi)
+        {
+          gvertices_data[vi] = pmesh->GroupVertex(gi, vi);
+        }
+      }
+
+      // For all these higher-order elements, we store a flat list of tuples of
+      // vertex indices instead of MFEM's internal edge/face indices for generality
+      // This uses more space but technically the Blueprint does not standardize how
+      // edge/face indices correspond to vertex indices
+      if(dim >= 2)
+      {
+        // Don't create the group if there are no elements
+        if(num_gedges > 0)
+        {
+          sidre::View* gedges_view =
+            group_grp->createViewAndAllocate("edges",
+                                             sidre::INT_ID,
+                                             num_gedges * 2);
+          int* gedges_data = gedges_view->getData<int*>();
+
+          for(int ei = 0; ei < num_gedges; ++ei)
+          {
+            pmesh->GroupEdge(gi, ei, tmp_edge, tmp_orientation);
+            pmesh->GetEdgeVertices(tmp_edge, tmp_verts);
+            // Copy into array such that vertices for a given edge are contiguous
+            // that is, v0 of e0, v1 of e0, v0 of e1, v1 of e1, v0 of e2, etc
+            std::copy(tmp_verts.begin(), tmp_verts.end(), &gedges_data[2 * ei]);
+          }
+        }
+
+        if(dim >= 3)
+        {
+          if(num_gtris > 0)
+          {
+            sidre::View* gtris_view =
+              group_grp->createViewAndAllocate("triangles",
+                                               sidre::INT_ID,
+                                               num_gtris * 3);
+            int* gtris_data = gtris_view->getData<int*>();
+            for(int ti = 0; ti < num_gtris; ++ti)
+            {
+              pmesh->GroupTriangle(gi, ti, tmp_face, tmp_orientation);
+              pmesh->GetFaceVertices(tmp_face, tmp_verts);
+              std::copy(tmp_verts.begin(), tmp_verts.end(), &gtris_data[3 * ti]);
+            }
+          }
+
+          if(num_gquads > 0)
+          {
+            sidre::View* gquads_view =
+              group_grp->createViewAndAllocate("quadrilaterals",
+                                               sidre::INT_ID,
+                                               num_gquads * 4);
+            int* gquads_data = gquads_view->getData<int*>();
+            for(int qi = 0; qi < num_gquads; ++qi)
+            {
+              pmesh->GroupQuadrilateral(gi, qi, tmp_face, tmp_orientation);
+              pmesh->GetFaceVertices(tmp_face, tmp_verts);
+              std::copy(tmp_verts.begin(), tmp_verts.end(), &gquads_data[4 * qi]);
+            }
+          }
+        }
       }
     }
   }
@@ -568,9 +656,8 @@ bool MFEMSidreDataCollection::verifyMeshBlueprint()
   bool result = conduit::blueprint::mesh::verify(mesh_node, verify_info);
   // conduit::Node::to_string only available in latest version
   SLIC_WARNING_IF(!result,
-                  "MFEMSidreDataCollection blueprint verification failed: " /*<<
-                                                                               verify_info.to_string()
-                                                                               */                          );
+                  "MFEMSidreDataCollection blueprint verification failed: "
+                  /*<< verify_info.to_string()*/);
   return result;
 }
 
@@ -738,11 +825,22 @@ void MFEMSidreDataCollection::Load(const std::string& path,
   // variables.
   if(m_owns_datastore)
   {
-    SetGroupPointers(m_datastore_ptr->getRoot()->getGroup(name + "_global"),
+    // Use the same path format as was used to create the datastore
+    SetGroupPointers(m_datastore_ptr->getRoot()->getGroup(
+                       name + "_global/blueprint_index/" + name),
                      m_datastore_ptr->getRoot()->getGroup(name));
+    SLIC_ERROR_IF(m_bp_grp->getNumGroups() == 0,
+                  "Loaded datastore is empty, was the datastore created on a "
+                  "different number of nodes?");
 
     UpdateStateFromDS();
   }
+
+  // Create a mesh from the datastore that was just read in
+  reconstructMesh();
+
+  // Create any fields from the datastore that was just read in
+  reconstructFields();
 }
 
 void MFEMSidreDataCollection::LoadExternalData(const std::string& path)
@@ -1068,7 +1166,8 @@ void MFEMSidreDataCollection::RegisterField(const std::string& field_name,
     }
   }
 
-  sidre::Group* grp = f->createGroup(field_name);
+  // This will return the existing field (if external), otherwise, a new group
+  sidre::Group* grp = alloc_group(f, field_name);
 
   // Set the "basis" string using the gf's finite element space, overwrite if
   // necessary.
@@ -1285,6 +1384,202 @@ std::string MFEMSidreDataCollection::getElementName(mfem::Element::Type elementE
   }
 
   return "unknown";
+}
+
+mfem::Geometry::Type MFEMSidreDataCollection::getElementTypeFromName(
+  const std::string& name)
+{
+  if(name == "point")
+    return mfem::Geometry::POINT;
+  else if(name == "line")
+    return mfem::Geometry::SEGMENT;
+  else if(name == "tri")
+    return mfem::Geometry::TRIANGLE;
+  else if(name == "quad")
+    return mfem::Geometry::SQUARE;
+  else if(name == "tet")
+    return mfem::Geometry::TETRAHEDRON;
+  else if(name == "hex")
+    return mfem::Geometry::CUBE;
+  else
+    return mfem::Geometry::INVALID;
+}
+
+// private method
+void MFEMSidreDataCollection::reconstructMesh()
+{
+  // mfem::ConduitDataCollection::BlueprintMeshToMesh would be useful here, but
+  // we need all the parameters to construct a ParMesh (the Mesh base subobject
+  // is initialized manually)
+
+  SLIC_ERROR_IF(
+    !verifyMeshBlueprint(),
+    "Cannot reconstruct mesh, data does not satisfy Conduit Blueprint");
+
+  SLIC_ERROR_IF(!m_bp_grp->hasView("coordsets/coords/values/x"),
+                "Cannot reconstruct a mesh without a Cartesian coordinate set");
+
+  View* vertex_view = m_bp_grp->getView("coordsets/coords/values/x");
+  SLIC_ERROR_IF(vertex_view->isExternal(),
+                "Cannot reconstruct a mesh that was built using external data");
+
+  // Assumes that Vertex is a standard layout type with only an array member
+  static_assert(std::is_standard_layout<mfem::Vertex>::value,
+                "mfem::Vertex must have standard layout to determine size of "
+                "its array member");
+  const std::size_t EXPECTED_STRIDE = sizeof(mfem::Vertex) / sizeof(double);
+  if((vertex_view->getTypeID() != DataTypeId::DOUBLE_ID) ||
+     (vertex_view->getStride() != EXPECTED_STRIDE))
+  {
+    SLIC_ERROR("Vertex array must consist of interleaved doubles");
+  }
+  // Sufficient to always grab the underlying data to the x-coords
+  // as the layout of the internal buffer is identical to what MFEM
+  // expects, i.e., x0 y0 z0 x1 y1 z1 regardless of dimension
+  double* vertices = vertex_view->getData();
+
+  // Use the x to get the number of vertices
+  int num_vertices = vertex_view->getNumElements();
+
+  SLIC_ERROR_IF(!m_bp_grp->hasGroup("topologies/mesh/elements"),
+                "Cannot reconstruct mesh without mesh topology");
+
+  int* element_indices =
+    m_bp_grp->getView("topologies/mesh/elements/connectivity")->getData<int*>();
+
+  // Name of the element type - convert to mfem::Geometry::Type later
+  std::string element_name =
+    m_bp_grp->getView("topologies/mesh/elements/shape")->getString();
+
+  View* element_attribute_view =
+    m_bp_grp->getView("fields/mesh_material_attribute/values");
+
+  int* element_attributes = element_attribute_view->getData<int*>();
+  int num_elements = element_attribute_view->getNumElements();
+
+  SLIC_ERROR_IF(!m_bp_grp->hasGroup("topologies/boundary/elements"),
+                "Cannot reconstruct mesh without boundary topology");
+
+  int* boundary_indices =
+    m_bp_grp->getView("topologies/boundary/elements/connectivity")->getData<int*>();
+
+  // Name of the element type - convert to mfem::Geometry::Type later
+  std::string bdr_element_name =
+    m_bp_grp->getView("topologies/boundary/elements/shape")->getString();
+
+  View* bdr_attribute_view =
+    m_bp_grp->getView("fields/boundary_material_attribute/values");
+
+  int* boundary_attributes = bdr_attribute_view->getData<int*>();
+  int num_boundary_elements = bdr_attribute_view->getNumElements();
+
+  int dimension = 1;
+
+  if(m_bp_grp->hasView("coordsets/coords/values/z"))
+  {
+    dimension = 3;
+  }
+
+  else if(m_bp_grp->hasView("coordsets/coords/values/y"))
+  {
+    dimension = 2;
+  }
+
+  #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+  // If it has an adjacencies group, the reloaded state was a ParMesh
+  if(m_bp_grp->hasGroup("adjsets/mesh"))
+  {
+    SLIC_ERROR(
+      "[MFEMSidreDataCollection]: Restoring parallel meshes is not yet "
+      "supported");
+  }
+  else
+  #endif
+  {
+    m_owned_mesh = std::unique_ptr<mfem::Mesh>(
+      new mfem::Mesh(vertices,
+                     num_vertices,
+                     element_indices,
+                     getElementTypeFromName(element_name),
+                     element_attributes,
+                     num_elements,
+                     boundary_indices,
+                     getElementTypeFromName(bdr_element_name),
+                     boundary_attributes,
+                     num_boundary_elements,
+                     dimension));
+  }
+  // Now that we've initialized an owning pointer, set the base subobject's
+  // mesh pointer as a non-owning pointer
+  mesh = m_owned_mesh.get();
+}
+
+void MFEMSidreDataCollection::reconstructFields()
+{
+  sidre::Group* f = m_bp_grp->getGroup("fields");
+  for(auto idx = f->getFirstValidGroupIndex(); sidre::indexIsValid(idx);
+      idx = f->getNextValidGroupIndex(idx))
+  {
+    Group* field_grp = f->getGroup(idx);
+    // Filter out the non-user-registered attribute fields
+    if(!field_grp->hasView("association"))
+    {
+      // FiniteElementCollection
+      auto basis_name = field_grp->getView("basis")->getString();
+      m_fecolls.emplace_back(mfem::FiniteElementCollection::New(basis_name));
+
+  // FiniteElementSpace - mesh ptr and FEColl ptr
+  #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+      auto parmesh = dynamic_cast<mfem::ParMesh*>(mesh);
+      if(parmesh)
+      {
+        m_fespaces.emplace_back(
+          new mfem::ParFiniteElementSpace(parmesh, m_fecolls.back().get()));
+      }
+      else
+  #endif
+      {
+        m_fespaces.emplace_back(
+          new mfem::FiniteElementSpace(mesh, m_fecolls.back().get()));
+      }
+
+      double* values = nullptr;
+      // Scalar grid function
+      if(field_grp->hasView("values"))
+      {
+        values = field_grp->getView("values")->getData();
+      }
+
+      // Vector grid function
+      else if(field_grp->hasGroup("values"))
+      {
+        // Sufficient to use address of first component as data is interleaved
+        values = field_grp->getGroup("values")->getView("x0")->getData();
+      }
+
+      else
+      {
+        SLIC_ERROR("Cannot reconstruct grid function - field values not found");
+      }
+  #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+      auto parfes =
+        dynamic_cast<mfem::ParFiniteElementSpace*>(m_fespaces.back().get());
+      if(parfes)
+      {
+        m_owned_gridfuncs.emplace_back(new mfem::ParGridFunction(parfes, values));
+      }
+      else
+  #endif
+      {
+        m_owned_gridfuncs.emplace_back(
+          new mfem::GridFunction(m_fespaces.back().get(), values));
+      }
+
+      // Register a non-owning pointer with the base subobject
+      DataCollection::RegisterField(field_grp->getName(),
+                                    m_owned_gridfuncs.back().get());
+    }
+  }
 }
 
 } /* namespace sidre */