+++ src/axom/sidre/core/MFEMSidreDataCollection.hpp
@@ -165,13 +165,9 @@ namespace sidre
 
     @note QuadratureFunction%s (q-fields) are not supported.
 
-    @note MFEMSidreDataCollection does not manage the FiniteElementSpace%s and
-    FiniteElementCollection%s associated with registered GridFunction%s.
-    Therefore, field registration is left to the user of MFEMSidreDataCollection
-       and
-    there are no methods that automatically register GridFunction%s using just
-    the content of the Sidre DataStore. Such capabilities can be implemented in
-    a derived class, adding any desired object management routines.
+    @note MFEMSidreDataCollection will attempt to reconstruct meshes and fields
+    on a restart, and will automatically register them.  This functionality is
+    experimental.
 
     @warning This class is still _experimental_, meaning that in future
     releases, it may not be backward compatible, and the output files generated
@@ -342,7 +338,7 @@ public:
   void Save(const std::string& filename, const std::string& protocol);
 
   /// Load the Sidre DataStore from file.
-  /** No mesh or fields are read from the loaded DataStore.
+  /** The mesh and fields will be read from the loaded datastore
 
       If the data collection created the datastore, it knows the layout of
       where the domain and global groups are, and can restore them after the
@@ -403,10 +399,13 @@ public:
                          IndexType sz,
                          TypeID type = DOUBLE_ID);
 
-  /// Deallocate the named buffer @a buffer_name.
+  /// Deallocate the named buffer @a buffer_name, if allocated.
   void FreeNamedBuffer(const std::string& buffer_name)
   {
-    named_buffers_grp()->destroyViewAndData(buffer_name);
+    if(named_buffers_grp()->hasView(buffer_name))
+    {
+      named_buffers_grp()->destroyViewAndData(buffer_name);
+    }
   }
 
   /// Verifies that the contents of the mesh blueprint data is valid.
@@ -457,6 +456,16 @@ private:
   // This is stored for convenience.
   Group* m_named_bufs_grp;
 
+  // Used to retain ownership of components of reconstructed Meshes and GridFuncs
+  // Instead of using flags to keep track of ownership between this class
+  // and the mfem::DataCollection subobject, always have the subobject
+  // retain a non-owning pointer and manage memory via unique_ptr in
+  // this class for consistency
+  std::unique_ptr<mfem::Mesh> m_owned_mesh;
+  std::vector<std::unique_ptr<mfem::FiniteElementCollection>> m_fecolls;
+  std::vector<std::unique_ptr<mfem::FiniteElementSpace>> m_fespaces;
+  std::vector<std::unique_ptr<mfem::GridFunction>> m_owned_gridfuncs;
+
   // Private helper functions
 
   void RegisterFieldInBPIndex(const std::string& field_name,
@@ -469,6 +478,19 @@ private:
   /** @brief Return a string with the conduit blueprint name for the given
       Element::Type. */
   std::string getElementName(mfem::Element::Type elementEnum);
+
+  /** @brief Return an mfem::Geometry::Type for the given
+      string with the conduit blueprint name. */
+  // Why is there mfem::Element::Type and mfem::Geometry::Type? They look the same
+  mfem::Geometry::Type getElementTypeFromName(const std::string& name);
+
+  // Reconstructs a mesh using the current contents of the datastore
+  // Used as part of Load()
+  void reconstructMesh();
+
+  // Reconstructs all non-mesh-related fields using the current contents
+  // of the datastore, used as part of Load()
+  void reconstructFields();
 
   /**
    * \brief A private helper function to set up the views associated with the