+++ src/axom/sidre/tests/sidre_mfem_datacollection.cpp
@@ -18,6 +18,7 @@ using axom::sidre::Group;
 using axom::sidre::MFEMSidreDataCollection;
 
 const std::string COLL_NAME = "test_collection";
+const double EPSILON = 1.0e-6;
 
 TEST(sidre_datacollection, dc_alloc_no_mesh)
 {
@@ -93,12 +94,12 @@ TEST(sidre_datacollection, dc_save)
   EXPECT_TRUE(sdc.verifyMeshBlueprint());
 }
 
-TEST(sidre_datacollection, dc_reload)
+TEST(sidre_datacollection, dc_reload_gf)
 {
   const std::string field_name = "test_field";
-  // 1D mesh divided into 10 segments
-  mfem::Mesh mesh(10);
-  mfem::H1_FECollection fec(1, 1);
+  // 2D mesh divided into triangles
+  mfem::Mesh mesh(10, 10, mfem::Element::TRIANGLE);
+  mfem::H1_FECollection fec(1, mesh.Dimension());
   mfem::FiniteElementSpace fes(&mesh, &fec);
 
   // The mesh and field(s) must be owned by Sidre to properly manage data in case of
@@ -110,24 +111,86 @@ TEST(sidre_datacollection, dc_reload)
   // Register to allocate storage internally, then write to it
   sdc_writer.RegisterField(field_name, &gf_write);
 
-  mfem::ConstantCoefficient one(1.0);
-  gf_write.ProjectCoefficient(one);
+  mfem::ConstantCoefficient three_and_a_half(3.5);
+  gf_write.ProjectCoefficient(three_and_a_half);
 
   EXPECT_TRUE(sdc_writer.verifyMeshBlueprint());
 
+  #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+  sdc_writer.SetComm(MPI_COMM_WORLD);
+  #endif
+
   sdc_writer.SetPrefixPath("/tmp/dc_reload_test");
   sdc_writer.SetCycle(0);
-  sdc_writer.PrepareToSave();
   sdc_writer.Save();
 
+  // No mesh is used here
   MFEMSidreDataCollection sdc_reader(COLL_NAME);
+
+  #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+  sdc_reader.SetComm(MPI_COMM_WORLD);
+  #endif
+
   sdc_reader.SetPrefixPath("/tmp/dc_reload_test");
   sdc_reader.Load();
+
+  // No need to reregister, it already exists
+  auto gf_read = sdc_reader.GetField(field_name);
+
+  // Make sure the gridfunction was actually read in
+  EXPECT_LT(gf_read->ComputeL2Error(three_and_a_half), EPSILON);
+
+  EXPECT_TRUE(sdc_reader.verifyMeshBlueprint());
+}
+
+TEST(sidre_datacollection, dc_reload_mesh)
+{
+  const std::string field_name = "test_field";
+  // 2D mesh divided into triangles
+  mfem::Mesh mesh(10, 10, mfem::Element::TRIANGLE);
+  mfem::H1_FECollection fec(1, mesh.Dimension());
+  mfem::FiniteElementSpace fes(&mesh, &fec);
+
+  // The mesh and field(s) must be owned by Sidre to properly manage data in case of
+  // a simulated restart (save -> load)
+  bool owns_mesh = true;
+  MFEMSidreDataCollection sdc_writer(COLL_NAME, &mesh, owns_mesh);
+
+  EXPECT_TRUE(sdc_writer.verifyMeshBlueprint());
+
+  // Save some basic info about the mesh
+  const int n_verts = sdc_writer.GetMesh()->GetNV();
+  const int n_ele = sdc_writer.GetMesh()->GetNE();
+  const int n_bdr_ele = sdc_writer.GetMesh()->GetNBE();
+
+  #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+  sdc_writer.SetComm(MPI_COMM_WORLD);
+  #endif
+
+  sdc_writer.SetPrefixPath("/tmp/dc_reload_test");
+  sdc_writer.SetCycle(0);
+  sdc_writer.Save();
+
+  // No mesh is used here
+  MFEMSidreDataCollection sdc_reader(COLL_NAME);
+
+  #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+  sdc_reader.SetComm(MPI_COMM_WORLD);
+  #endif
+
+  sdc_reader.SetPrefixPath("/tmp/dc_reload_test");
+  sdc_reader.Load();
+
+  // Make sure the mesh was actually reconstructed
+  EXPECT_EQ(sdc_reader.GetMesh()->GetNV(), n_verts);
+  EXPECT_EQ(sdc_reader.GetMesh()->GetNE(), n_ele);
+  EXPECT_EQ(sdc_reader.GetMesh()->GetNBE(), n_bdr_ele);
 
   EXPECT_TRUE(sdc_reader.verifyMeshBlueprint());
 }
 
   #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)
+
 TEST(sidre_datacollection, dc_alloc_owning_parmesh)
 {
   // 1D mesh divided into 10 segments
@@ -146,29 +209,6 @@ TEST(sidre_datacollection, dc_alloc_nonowning_parmesh)
   bool owns_mesh = false;
   MFEMSidreDataCollection sdc(COLL_NAME, &parmesh, owns_mesh);
   EXPECT_TRUE(sdc.verifyMeshBlueprint());
-}
-
-TEST(sidre_datacollection, dc_par_reload)
-{
-  // 1D mesh divided into 10 segments
-  mfem::Mesh mesh(10);
-  mfem::ParMesh parmesh(MPI_COMM_WORLD, mesh);
-  // The mesh must be owned by Sidre to properly manage data in case of
-  // a simulated restart (save -> load)
-  bool owns_mesh = true;
-  MFEMSidreDataCollection sdc_writer(COLL_NAME, &parmesh, owns_mesh);
-
-  sdc_writer.SetPrefixPath("/tmp/dc_par_reload_test");
-  sdc_writer.SetCycle(0);
-  sdc_writer.PrepareToSave();
-  sdc_writer.Save();
-
-  MFEMSidreDataCollection sdc_reader(COLL_NAME);
-  sdc_reader.SetComm(MPI_COMM_WORLD);
-  sdc_reader.SetPrefixPath("/tmp/dc_par_reload_test");
-  sdc_reader.Load();
-
-  EXPECT_TRUE(sdc_reader.verifyMeshBlueprint());
 }
 
     //----------------------------------------------------------------------